refactor: organize source files
